Make dev code removal work for bundled-es exports
Fixes #1063
