+++ lib/bundle/concat_source_amd.js
@@ -3,7 +3,10 @@ var concat = require("../source-map-concat");
 var sourceNode = require("../node/source").node;
 var removeSourceMapUrl = require("../remove_source_map_url");
 
-module.exports = function(bundle, sourceProp, excludePlugins){
+module.exports = function(bundle, options){
+	var sourceProp = options.sourceProp;
+	var excludePlugins = options.excludePlugins;
+	
 	var output = fileName(bundle);
 
 	var makeCode = function(name) {
