+++ lib/bundle/concat_source_es.js
@@ -1,12 +1,16 @@
+const cleanSource = require("../node/clean");
 const dependencyResolver = require("../node/dependency_resolver");
 const isProcessShim = require("../node/is_process_shim");
+const nodeSource = require("../node/source");
 const pluginCommonjs = require("rollup-plugin-commonjs");
+const removeActiveSourceKeys = require("../node/remove_active_source_keys");
 const rollup = require("steal-rollup");
 
 const moduleNameFromSpecifier = dependencyResolver.moduleNameFromSpecifier;
 const CJS_PROXY_PREFIX = '\0commonjs-proxy:';
 
-module.exports = function(bundle) {
+module.exports = function(bundle, options) {
+	let removeDevelopmentCode = options.removeDevelopmentCode;
 	let main = bundle.bundles[0];
 	let firstNode = bundle.nodes[0];
 
@@ -20,7 +24,7 @@ module.exports = function(bundle) {
 		input: main,
 
 		plugins: [
-			loadFromGraph(getNode),
+			loadFromGraph(getNode, removeDevelopmentCode),
 			pluginCommonjs({})
 		]
 	})
@@ -43,7 +47,7 @@ module.exports = function(bundle) {
 	});
 };
 
-function loadFromGraph(getNode) {
+function loadFromGraph(getNode, removeDevelopmentCode) {
 	return {
 		resolveId: function(id, importer) {
 			if(id.startsWith(CJS_PROXY_PREFIX)) {
@@ -68,8 +72,13 @@ function loadFromGraph(getNode) {
 				return undefined;
 			}
 
-			let source = node.load.source;
-			return source;
+			removeActiveSourceKeys(node);
+
+			if(removeDevelopmentCode) {
+				cleanSource(node, {});
+			}
+
+			return nodeSource(node);
 		}
 	};
 }
