+++ lib/graph/remove_active_source_keys.js
@@ -1,12 +1,4 @@
-var clean = function(node, options){
-	delete node.activeSourceKeys;
-	node.activeSource = {
-		code: node.load.source
-	};
-	if(options.removeSourceNodes) {
-		delete node.sourceNode;
-	}
-};
+var clean = require("../node/remove_active_source_keys");
 
 module.exports = function(graph, options) {
 	options = options || {};
@@ -22,4 +14,3 @@ module.exports = function(graph, options) {
 		}
 	}
 };
-
