+++ lib/node/remove_active_source_keys.js
@@ -1,0 +1,10 @@
+
+module.exports = function(node, options = {}){
+	delete node.activeSourceKeys;
+	node.activeSource = {
+		code: node.load.source
+	};
+	if(options.removeSourceNodes) {
+		delete node.sourceNode;
+	}
+};
