+++ test/export_bundled_es_test.js
@@ -1,4 +1,5 @@
 var assert = require("assert");
+var fs = require("fs");
 var stealExport = require("../lib/build/export");
 var rmdir = require("rimraf");
 var testHelpers = require("./helpers");
@@ -71,5 +72,27 @@ describe("+bundled-es", function(){
 			assert.ok(true, "it built");
 			done();
 		}, done);
+	});
+
+	it("Removes dev code", function(done){
+		this.timeout(10000);
+		stealExport({
+			steal: {
+				config: __dirname+"/exports_basics/package.json!npm",
+				main: "app/index_es"
+			},
+			options: { quiet: true },
+			"outputs": {
+				"+bundled-es": {
+					dest: __dirname + "/exports_basics/dist/es.js"
+				}
+			}
+		})
+		.then(function() {
+			var code = fs.readFileSync(__dirname + "/exports_basics/dist/es.js", "utf8");
+			assert.ok(!/window\.REMOVEME/.test(code), "Removed the dev code");
+			done();
+		})
+		.catch(done);
 	});
 });
