+++ test/exports_basics/index_es.js
@@ -2,6 +2,10 @@ import foo from "foo";
 
 let name = "index";
 
+//!steal-remove-start
+window.REMOVEME = true;
+//!steal-remove-end
+
 export {
 	name,
 	foo
